FDR pipeline updates
- Ported the new "falcon" functionality to the "fdr" pipeline
- Refactored and added tests
- Simplified the pipeline code.
- Removed some unused/unreached code from the logscale backend
